requirements.txt
asgiref==3.8.1
Django==4.2.17
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
et_xmlfile==2.0.0
openpyxl==3.1.5
sqlparse==0.5.3
typing_extensions==4.12.2
